refactor(lib::podcast): extract "Episode" structs and impls into own module
